backend/requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
asgi-lifespan==2.1.0
attrs==23.1.0
bcrypt==4.0.1
black==23.9.1
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
coverage==6.5.0
coveralls==3.3.1
cryptography==41.0.4
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
Faker==20.1.0
flake8==6.1.0
gevent==23.9.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
mailjet-rest==1.3.4
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.0
orjson==3.9.7
packaging==23.1
pandas==2.1.0
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==0.13.1
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
pytest==6.2.4
pytest-asyncio==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
zope.event==5.0
zope.interface==6.0
lxml==4.9.3
